requirements.txt
alembic==1.12.0
apispec==6.3.0
blinker==1.6.2
click==8.1.7
dnspython==2.4.2
email-validator==2.0.0.post2
Flask==3.0.0
Flask-Login==0.6.2
Flask-Migrate==4.0.5
flask-smorest==0.42.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-enum==1.5.1
packaging==23.2
psycopg2-binary==2.9.9
python-slugify==8.0.1
SQLAlchemy==2.0.21
text-unidecode==1.3
typing_extensions==4.8.0
webargs==8.3.0
Werkzeug==1.0.1
WTForms==3.0.1
